requirements.txt
wagtail
requests
python-dotenv
django-click
